Moving tests into new place
