+++ graylog-storage-elasticsearch6/src/test/java/org/graylog/storage/elasticsearch6/views/IndexLookupTest.java
@@ -14,8 +14,9 @@
  * You should have received a copy of the GNU General Public License
  * along with Graylog.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.graylog.plugins.views.search.elasticsearch;
+package org.graylog.storage.elasticsearch6.views;
 
+import org.graylog.plugins.views.search.elasticsearch.IndexLookup;
 import org.graylog2.indexer.ranges.IndexRange;
 import org.graylog2.indexer.ranges.IndexRangeService;
 import org.graylog2.plugin.database.Persisted;
